Enable postprocessing of article HTML
... right before the article is written to the disk.

Add `:postprocess-article-html-fn` config option that will be called on the article
(page/post) after it is transformed from markup to HTML.

This can be used e.g. to run it through Selmer so that you can use Selmer templating
in your markup and get it processed - see https://github.com/cryogen-project/cryogen-core/pull/139 for the original request for that.