Fail if extend-params-fn adds :posts
because 1) it would break rendering of posts and pages due to the order
of the cond clauses in `htmlize-content`
2) even if we changed the order so that posts is last (which seems to
work fine) it would still not work as expected because some pages
define :posts themselves, overriding this - e.g. tag pages.

The best solution is thus tell the user not to do this and use another key instead.

Fixes #156