Fixed URL encoding in page-uri for all URIs.
After a discussion with yogthos, decided it was better to make the URL
encoding change in `page-uri` for all URIs, rather than piecemeal for
each particular case.

Had to create a new function called `page-file-name` which has similar
functionality as `page-uri` but does not do URL encoding.  This is
because `page-uri` was used in `tag-info` to create a file path and a
uri.  I believe `tag-info` is the only place that needs
`page-file-name`, but not sure.

Added a test to check for proper encoding of author URLs.